feat: Add videoMain component and update spin-tick animation duration
The code changes add the videoMain component to the mainpage directory, which displays a video with auto play, loop, and muted attributes. This enhances the visual experience on the main page. Additionally, the spin-tick animation duration is updated to 30 seconds with 30 steps in the tailwind.config.ts file. and added scroll feature and updateed responsiveness